Mass Refactoring. Created Abstract Factory
