--- src/com/github/igotyou/FactoryMod/interfaces/Properties.java
@@ -1,24 +1,0 @@
-package com.github.igotyou.FactoryMod.interfaces;
-
-import com.github.igotyou.FactoryMod.utility.ItemList;
-import com.github.igotyou.FactoryMod.utility.NamedItemStack;
-
-//original file:
-/**
- * Properties.java
- * Purpose: Interface for Properties objects for basic properties functionality
- *
- * @author MrTwiggy
- * @version 0.1 1/17/13
- */
-//edited version:
-/**
-* Properties.java	 
- * Purpose: Interface for Properties objects for basic properties functionality
-* @author igotyou
-*
-*/
-public interface Properties 
-{
-	public ItemList<NamedItemStack> getInputs() ;
-}
