+++ src/com/github/igotyou/FactoryMod/persistence/FactoryCodec.java
@@ -1,8 +1,8 @@
 package com.github.igotyou.FactoryMod.persistence;
 
-import com.github.igotyou.FactoryMod.interfaces.Factory;
+import com.github.igotyou.FactoryMod.Factorys.IFactory;
 
-public interface FactoryCodec<T extends Factory> {
+public interface FactoryCodec<T extends IFactory> {
 	
 	public void write(T out);
