+++ src/com/github/igotyou/FactoryMod/persistence/FactoryReader.java
@@ -2,9 +2,9 @@ package com.github.igotyou.FactoryMod.persistence;
 
 import java.util.List;
 
-import com.github.igotyou.FactoryMod.interfaces.Factory;
+import com.github.igotyou.FactoryMod.Factorys.IFactory;
 
-public interface FactoryReader<T extends Factory> {
+public interface FactoryReader<T extends IFactory> {
 	
 	public List<T> read();
