+++ src/com/github/igotyou/FactoryMod/persistence/FactoryWriter.java
@@ -2,9 +2,9 @@ package com.github.igotyou.FactoryMod.persistence;
 
 import java.util.List;
 
-import com.github.igotyou.FactoryMod.interfaces.Factory;
+import com.github.igotyou.FactoryMod.Factorys.IFactory;
 
-public interface FactoryWriter<T extends Factory> {
+public interface FactoryWriter<T extends IFactory> {
 	
 	public void write(List<T> factories);
