+++ src/com/github/igotyou/FactoryMod/properties/AFactoryProperties.java
@@ -1,0 +1,10 @@
+package com.github.igotyou.FactoryMod.properties;
+
+public abstract class AFactoryProperties implements IFactoryProperties{
+	protected String name;
+	
+	public String getName()
+	{
+		return name;
+	}
+}
