+++ src/com/github/igotyou/FactoryMod/properties/IFactoryProperties.java
@@ -1,0 +1,25 @@
+package com.github.igotyou.FactoryMod.properties;
+
+import com.github.igotyou.FactoryMod.utility.ItemList;
+import com.github.igotyou.FactoryMod.utility.NamedItemStack;
+
+//original file:
+/**
+ * Properties.java
+ * Purpose: Interface for Properties objects for basic properties functionality
+ *
+ * @author MrTwiggy
+ * @version 0.1 1/17/13
+ */
+//edited version:
+/**
+* Properties.java	 
+ * Purpose: Interface for Properties objects for basic properties functionality
+* @author igotyou
+*
+*/
+public interface IFactoryProperties 
+{
+//	public ItemList<NamedItemStack> getInputs() ;
+	public String getName();
+}
